+++ tests/testthat/test-keyATMCovDirMulti.R
@@ -111,7 +111,7 @@ test_that("Covariates settings: Standardize - none", {
 
   skip_on_os("linux") ; skip_on_cran()
   expect_error(predict(cov, bills_cov_modified))
-  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[3, 3]), 0.000084265, tolerance = 0.0001)
+  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[3, 3]), 0.000084265, tolerance = 0.003)
 })