+++ NEWS.md
@@ -2,6 +2,7 @@
 ### Major changes
 * Migrating to C++17 to follow [the new CRAN check](https://developer.r-project.org/blosxom.cgi/R-devel/2023/01/27#n2023-01-27). We edited `shuffled_indexes()` that internally used `std::random_shuffle()`. This change does not guarantee backward compatibility across all platforms.
 * A bug fix in `summary.keyATM_docs()`.
+* The default number of iterations is `3000` (previously `1500`). This is the number of iterations we used in our paper.
 
 # keyATM 0.4.2