+++ NEWS.md
@@ -1,3 +1,10 @@
+# keyATM 0.5.1
+
+### Minor changes
+* Adding options to set the hyperparameters `eta`.
+* Documentation updates to pass CRAN checks.
+
+
 # keyATM 0.5.0
 ### Major changes
 * Migrating to C++17 to follow [the new CRAN check](https://developer.r-project.org/blosxom.cgi/R-devel/2023/01/27#n2023-01-27). We edited `shuffled_indexes()` that internally used `std::random_shuffle()`. This change does not guarantee backward compatibility across all platforms.